Problem: Duplicate defintions.
Solution: Move them to fty-common-db, remove unnecessary file.
