+++ project.xml
@@ -98,7 +98,6 @@
         <use project = "openssl" header = "openssl/sha.h" debian_name="libssl-dev"/>
     </use>
 
-    <class name = "defs" state="stable" private = "1"> Defines constants </class>
     <class name = "dbhelpers" state="stable" private = "1"> Helpers functions for database </class>
     <class name = "fty_asset_uptime_configurator" state = "stable" private = "1">Configuration for kpi-uptime </class>
