+++ src/dbhelpers.h
@@ -35,23 +35,6 @@
 #include <string>
 #include "fty_asset_classes.h"
 
-//#include "preproc.h"
-
-// all fields called name
-#define MAX_NAME_LENGTH         50
-// t_bios_asset_ext_attributes.keytag
-#define MAX_KEYTAG_LENGTH       40
-// t_bios_asset_ext_attributes.value
-#define MAX_VALUE_LENGTH        255
-// t_bios_asset_device.mac
-#define MAX_MAC_LENGTH          17
-// t_bios_asset_device.hostname
-#define MAX_HOSTNAME_LENGTH     25
-// t_bios_asset_device.fullhostname
-#define MAX_FULLHOSTNAME_LENGTH 45
-
-#define MAX_DESCRIPTION_LENGTH  255
-
 // Selects all assets which have our container in input power chain
 FTY_ASSET_PRIVATE int
     select_links_by_container (
