--- src/defs.cc
@@ -1,43 +1,0 @@
-/*  =========================================================================
-    defs - Defines constants
-
-    Copyright (C) 2014 - 2015 Eaton                                        
-                                                                           
-    This program is free software; you can redistribute it and/or modify   
-    it under the terms of the GNU General Public License as published by   
-    the Free Software Foundation; either version 2 of the License, or      
-    (at your option) any later version.                                    
-                                                                           
-    This program is distributed in the hope that it will be useful,        
-    but WITHOUT ANY WARRANTY; without even the implied warranty of         
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the          
-    GNU General Public License for more details.                           
-                                                                           
-    You should have received a copy of the GNU General Public License along
-    with this program; if not, write to the Free Software Foundation, Inc.,
-    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.            
-    =========================================================================
-*/
-
-/*
-@header
-    defs - Defines constants
-@discuss
-@end
-*/
-
-#include "fty_asset_classes.h"
-
-
-//  --------------------------------------------------------------------------
-//  Self test of this class
-
-void
-defs_test (bool verbose)
-{
-    printf (" * defs: ");
-
-
-    //  @end
-    printf ("OK\n");
-}
